feat: Org unit TEA filter in enrollment aggregate [DHIS2-19938]
