+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/event/data/JdbcEnrollmentAnalyticsManager.java
@@ -31,7 +31,6 @@ package org.hisp.dhis.analytics.event.data;
 
 import static java.util.stream.Collectors.joining;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
-import static org.apache.commons.lang3.StringUtils.substringBetween;
 import static org.hisp.dhis.analytics.AnalyticsConstants.ANALYTICS_TBL_ALIAS;
 import static org.hisp.dhis.analytics.DataType.BOOLEAN;
 import static org.hisp.dhis.analytics.common.CteContext.ENROLLMENT_AGGR_BASE;
@@ -554,37 +553,6 @@ public class JdbcEnrollmentAnalyticsManager extends AbstractJdbcEventAnalyticsMa
     }
 
     return "select " + StringUtils.join(selectCols, ",") + " ";
-  }
-
-  /**
-   * This method switches or add a new prefix to the given column if needed. It takes into
-   * consideration columns used in functions as well as regular columns and columns with aliases.
-   *
-   * <p>ie:
-   *
-   * <ul>
-   *   <li>ax.value as "A03MvHHogjR" -> ax.value as "A03MvHHogjR"
-   *   <li>ev.value as "A03MvHHogjR" -> ev.value as "A03MvHHogjR"
-   *   <li>value as "A03MvHHogjR" -> ax.value as "A03MvHHogjR"
-   *   <li>count() as value -> count() as value"
-   *   <li>ST_Y(ax.geometry) -> ST_Y(ax.geometry)
-   * </ul>
-   *
-   * @param column to be prefixed.
-   * @return the prefixed column (if required).
-   */
-  String addEnrollmentPrefix(String column) {
-    String functionColumn = substringBetween(column, "(", ")");
-    boolean hasFunction = functionColumn != null;
-    boolean hasPrefix = column.contains("ax.") || column.contains("ev.");
-
-    if (!hasFunction && !hasPrefix) {
-      column = "ax." + column;
-    } else if (hasFunction && functionColumn.length() > 0 && !hasPrefix) {
-      column = column.replace(functionColumn, "ax." + functionColumn);
-    }
-
-    return column;
   }
 
   /**
